+++ Library/include/TextIO/StringSetIO.h
@@ -21,21 +21,21 @@ extern "C" {
      @param            NumStrings                        How many strings will this StringSet contain?
      @return                                             Returns an empty StringSet with room for NumStrings pointers plus a null terminator.
      */
-    UTF8             **UTF8_StringSet_Init(uint64_t NumStrings);
+    UTF8             **UTF8_StringSet_Init(size_t NumStrings);
 
     /*!
      @abstract                                           Creates a UTF-16 encoded StringSet.
      @param            NumStrings                        How many strings will this StringSet contain?
      @return                                             Returns an empty StringSet with room for NumStrings pointers plus a null terminator.
      */
-    UTF16            **UTF16_StringSet_Init(uint64_t NumStrings);
+    UTF16            **UTF16_StringSet_Init(size_t NumStrings);
 
     /*!
      @abstract                                           Creates a UTF-32 encoded StringSet.
      @param            NumStrings                        How many strings will this StringSet contain?
      @return                                             Returns an empty StringSet with room for NumStrings pointers plus a null terminator.
      */
-    UTF32            **UTF32_StringSet_Init(uint64_t NumStrings);
+    UTF32            **UTF32_StringSet_Init(size_t NumStrings);
 
     /*!
      @abstract                                           Attaches a string to a StringSet at the specified position.
@@ -44,7 +44,7 @@ extern "C" {
      @param            Index                             Which position should String2Attach be in?
      @return                                             Returns true when attaching was sucessful.
      */
-    bool               UTF8_StringSet_Attach(UTF8 **StringSet, UTF8 *String2Attach, uint64_t Index);
+    bool               UTF8_StringSet_Attach(UTF8 **StringSet, UTF8 *String2Attach, size_t Index);
 
     /*!
      @abstract                                           Attaches a string to a StringSet at the specified position.
@@ -53,7 +53,7 @@ extern "C" {
      @param            Index                             Which position should String2Attach be in?
      @return                                             Returns true when attaching was sucessful.
      */
-    bool               UTF16_StringSet_Attach(UTF16 **StringSet, UTF16 *String2Attach, uint64_t Index);
+    bool               UTF16_StringSet_Attach(UTF16 **StringSet, UTF16 *String2Attach, size_t Index);
 
     /*!
      @abstract                                           Attaches a string to a StringSet at the specified position.
@@ -62,63 +62,63 @@ extern "C" {
      @param            Index                             Which position should String2Attach be in?
      @return                                             Returns true when attaching was sucessful.
      */
-    bool               UTF32_StringSet_Attach(UTF32 **StringSet, UTF32 *String2Attach, uint64_t Index);
+    bool               UTF32_StringSet_Attach(UTF32 **StringSet, UTF32 *String2Attach, size_t Index);
 
     /*!
      @abstract                                           Gets the number of strings in a StringSet.
      @param            StringSet                         The StringSet to get the number of strings in.
      @return                                             Returns the number of strings in StringSet.
      */
-    uint64_t           UTF8_StringSet_GetNumStrings(ImmutableStringSet_UTF8 StringSet);
+    size_t             UTF8_StringSet_GetNumStrings(ImmutableStringSet_UTF8 StringSet);
 
     /*!
      @abstract                                           Gets the number of strings in a StringSet.
      @param            StringSet                         The StringSet to get the number of strings in.
      @return                                             Returns the number of strings in StringSet.
      */
-    uint64_t           UTF16_StringSet_GetNumStrings(ImmutableStringSet_UTF16 StringSet);
+    size_t             UTF16_StringSet_GetNumStrings(ImmutableStringSet_UTF16 StringSet);
 
     /*!
      @abstract                                           Gets the number of strings in a StringSet.
      @param            StringSet                         The StringSet to get the number of strings in.
      @return                                             Returns the number of strings in StringSet.
      */
-    uint64_t           UTF32_StringSet_GetNumStrings(ImmutableStringSet_UTF32 StringSet);
+    size_t             UTF32_StringSet_GetNumStrings(ImmutableStringSet_UTF32 StringSet);
 
     /*!
      @abstract                                           Gets the size of each string in the StringSet in code units.
      @param            StringSet                         The StringSet to get the size of each string.
      @return                                             Returns an array containing NumStrings elements, where each element contains the size.
      */
-    uint64_t          *UTF8_StringSet_GetStringSizesInCodeUnits(ImmutableStringSet_UTF8 StringSet);
+    size_t            *UTF8_StringSet_GetStringSizesInCodeUnits(ImmutableStringSet_UTF8 StringSet);
 
     /*!
      @abstract                                           Gets the size of each string in the StringSet in code units.
      @param            StringSet                         The StringSet to get the size of each string.
      @return                                             Returns an array containing NumStrings elements, where each element contains the size.
      */
-    uint64_t          *UTF16_StringSet_GetStringSizesInCodeUnits(ImmutableStringSet_UTF16 StringSet);
+    size_t            *UTF16_StringSet_GetStringSizesInCodeUnits(ImmutableStringSet_UTF16 StringSet);
 
     /*!
      @abstract                                           Gets the size of each string in the StringSet in code points.
      @param            StringSet                         The StringSet to get the size of each string.
      @return                                             Returns an array containing NumStrings elements, where each element contains the size.
      */
-    uint64_t          *UTF8_StringSet_GetStringSizesInCodePoints(ImmutableStringSet_UTF8 StringSet);
+    size_t            *UTF8_StringSet_GetStringSizesInCodePoints(ImmutableStringSet_UTF8 StringSet);
 
     /*!
      @abstract                                           Gets the size of each string in the StringSet in code points.
      @param            StringSet                         The StringSet to get the size of each string.
      @return                                             Returns an array containing NumStrings elements, where each element contains the size.
      */
-    uint64_t          *UTF16_StringSet_GetStringSizesInCodePoints(ImmutableStringSet_UTF16 StringSet);
+    size_t            *UTF16_StringSet_GetStringSizesInCodePoints(ImmutableStringSet_UTF16 StringSet);
 
     /*!
      @abstract                                           Gets the size of each string in the StringSet in code points.
      @param            StringSet                         The StringSet to get the size of each string.
      @return                                             Returns an array containing NumStrings elements, where each element contains the size.
      */
-    uint64_t          *UTF32_StringSet_GetStringSizesInCodePoints(ImmutableStringSet_UTF32 StringSet);
+    size_t            *UTF32_StringSet_GetStringSizesInCodePoints(ImmutableStringSet_UTF32 StringSet);
 
     /*!
      @abstract                                           Decodes a StringSet to a UTF32_StringSet.
